requirements.txt
mecab-python3==0.996.1
pykakasi==0.94
gokulang==1.0.2
falcon==1.4.1
